requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-ollama
faiss-cpu
pypdf  
sentence-transformers
langchain_huggingface